hub/main.py
import logging
from typing import List

import paho.mqtt.client as mqtt
from fastapi import FastAPI
from redis import Redis

from app.adapters.store_api_adapter import StoreApiAdapter
from app.entities.processed_agent_data import ProcessedAgentData
from config import (
    BATCH_SIZE,
    MQTT_BROKER_HOST,
    MQTT_BROKER_PORT,
    MQTT_TOPIC,
    REDIS_HOST,
    REDIS_PORT,
    STORE_API_BASE_URL,
)

logging.basicConfig(
    level=logging.INFO,
    format="[%(asctime)s] [%(levelname)s] [%(module)s] %(message)s",
    handlers=[
        logging.StreamHandler(),
        logging.FileHandler("app.log"),
    ],
)

redis_client = Redis(host=REDIS_HOST, port=REDIS_PORT)
store_adapter = StoreApiAdapter(api_base_url=STORE_API_BASE_URL)
app = FastAPI()


@app.post("/processed_agent_data/")
async def save_processed_agent_data(processed_agent_data: ProcessedAgentData):
    print(processed_agent_data)
    redis_client.lpush(
        "processed_agent_data",
        processed_agent_data.model_dump_json(),
    )

    processed_agent_data_batch: List[ProcessedAgentData] = []
    if redis_client.llen("processed_agent_data") >= BATCH_SIZE:
        for _ in range(BATCH_SIZE):
            processed_agent_data = ProcessedAgentData.model_validate_json(
                redis_client.lpop("processed_agent_data")
            )
            processed_agent_data_batch.append(processed_agent_data)

        store_adapter.save_data(
            processed_agent_data_batch=processed_agent_data_batch,
        )

    return {"status": "ok"}


client = mqtt.Client()


def on_connect(client, userdata, flags, rc):
    if rc == 0:
        logging.info("Connected to MQTT broker")
        client.subscribe(MQTT_TOPIC)
    else:
        logging.info(f"Failed to connect to MQTT broker with code: {rc}")


def on_message(client, userdata, msg):
    try:
        payload: str = msg.payload.decode("utf-8")
        processed_agent_data = ProcessedAgentData.model_validate_json(
            payload, strict=True
        )
        redis_client.lpush(
            "processed_agent_data",
            processed_agent_data.model_dump_json(),
        )

        processed_agent_data_batch: List[ProcessedAgentData] = []
        if redis_client.llen("processed_agent_data") >= BATCH_SIZE:
            for _ in range(BATCH_SIZE):
                processed_agent_data = ProcessedAgentData.model_validate_json(
                    redis_client.lpop("processed_agent_data")
                )
                processed_agent_data_batch.append(processed_agent_data)

            store_adapter.save_data(
                processed_agent_data_batch=processed_agent_data_batch,
            )
    except Exception as e:
        logging.info(f"Error processing MQTT message: {e}")


client.on_connect = on_connect
client.on_message = on_message
client.connect(MQTT_BROKER_HOST, MQTT_BROKER_PORT)
client.loop_start()
